Add support for generating Gatling clients
This set of templates generates clients using scala language, but with code
aimed at usage in Gatling. With this, instead of:

```
  http("get inventory")
    .get("/store/inventory")
    .queryParamsSeq(...)
    .headers(...)
    .body(...)
    .check(bodyString).saveAs("raw_string")
```

You can do:

```
  storeApi
    .getInventory(...)
    .check(StoreApi.getInventoryResponse()).saveAs("parsed_model")
```

Many things need to be done still, since they were missing in the scala
templates I based this on. For example enumerations are not supported.

Authentication and configuration isn't done the way it's handled in python or
ruby clients, so it requires more manual work.

And likely there are other things. But I thought I'll share it with the
community, hoping others will have time to contribute.